requirements.txt
streamlit==1.39.0
opencv-python-headless==4.7.0.72
numpy==1.26.4
pillow
fpdf==1.7.2
reportlab==4.2.0
pdf2image

